Se cambia de f32 a f64

+++ src/read_file.rs
@@ -8,14 +8,14 @@ fn read_lines(path:&str)-> Vec<String>{
         .collect()  // gather them together into a vector
 }
 
-pub fn listed_file_content(path:&str)->(Vec<char>, Vec<f32>){
+pub fn listed_file_content(path:&str)->(Vec<char>, Vec<f64>){
     let lines: Vec<String> = read_lines(path);
     let mut ascii: Vec<char> = vec![];
-    let mut brightness: Vec<f32> = vec![];
+    let mut brightness: Vec<f64> = vec![];
     for text in lines {
         let fisrs_char = &text[..1];
         let number = &text[2..];
-        let number = number.parse::<f32>().unwrap();
+        let number = number.parse::<f64>().unwrap();
         let val = fisrs_char.chars().take(1).last().unwrap();
         println!("{val}: {number}");
         ascii.push(val);
